pyproject.toml
[build-system]
requires = ["hatchling", "hatch-requirements-txt"]
build-backend = "hatchling.build"

[project]
name = "llmk-recipes"
version = "0.0.1"
authors = [
  { name="Hamid Shojanazeri", email="hamidnazeri@khulnasoft.com" },
  { name="Matthias Reso", email="mreso@khulnasoft.com" },
  { name="Geeta Chauhan", email="gchauhan@khulnasoft.com" },
]
description = "Llmk-recipes is a companion project to the Llmk 2 model. It's goal is to provide examples to quickly get started with fine-tuning for domain adaptation and how to run inference for the fine-tuned models. "
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
]
dynamic = ["dependencies"]

[project.optional-dependencies]
vllm = ["vllm"]
tests = ["pytest-mock"]
auditnlg = ["auditnlg"]

[project.urls]
"Homepage" = "https://github.com/khulnasoft/llmk-recipes/"
"Bug Tracker" = "https://github.com/khulnasoft/llmk-recipes/issues"

[tool.hatch.build]
exclude = [
  "dist/*",
]

[tool.hatch.build.targets.wheel]
packages = ["src/llmk_recipes"]

[tool.hatch.khulnasoftdata.hooks.requirements_txt]
files = ["requirements.txt"]

[tool.pytest.ini_options]
markers = [
    "skip_missing_tokenizer: skip tests when we can not access khulnasoft-llmk/Llmk-2-7b-hf on huggingface hub (Log in with `huggingface-cli login` to unskip).",
]
